+++ test/controller/alh_pdf_view_controller_test.dart
@@ -39,7 +39,8 @@ void main() {
   });
 
   tearDown(() {
-    channel.setMockMethodCallHandler(null);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, null);
   });
 
   group('#invoking methods', () {
@@ -217,11 +218,13 @@ void main() {
       const givenPage = 5;
       const givenSize = Size(100.0, 200.0);
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'pageSize') {
           actualMethodCall = call;
           return {'height': givenSize.height, 'width': givenSize.width};
         }
+        return null;
       });
 
       // when
@@ -244,11 +247,13 @@ void main() {
       // given
       const givenPageCount = 5;
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'pageCount') {
           actualMethodCall = call;
           return givenPageCount;
         }
+        return null;
       });
 
       // when
@@ -271,11 +276,13 @@ void main() {
       // given
       const givenPage = 5;
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'currentPage') {
           actualMethodCall = call;
           return givenPage;
         }
+        return null;
       });
 
       // when
@@ -297,11 +304,13 @@ void main() {
       // given
       const givenPage = 100;
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'setPage') {
           actualMethodCall = call;
           return true;
         }
+        return null;
       });
 
       // when
@@ -324,11 +333,13 @@ void main() {
       // given
       const givenPage = 100;
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'setPage') {
           actualMethodCall = call;
           return false;
         }
+        return null;
       });
 
       // when
@@ -352,11 +363,13 @@ void main() {
         'THEN should should call "nextPage" with expected arguments', () async {
       // given
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'nextPage') {
           actualMethodCall = call;
           return true;
         }
+        return null;
       });
 
       // when
@@ -377,11 +390,13 @@ void main() {
         'THEN should should call "nextPage" with expected arguments', () async {
       // given
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'nextPage') {
           actualMethodCall = call;
           return false;
         }
+        return null;
       });
 
       // when
@@ -405,11 +420,13 @@ void main() {
         () async {
       // given
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'previousPage') {
           actualMethodCall = call;
           return true;
         }
+        return null;
       });
 
       // when
@@ -431,11 +448,13 @@ void main() {
         () async {
       // given
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'previousPage') {
           actualMethodCall = call;
           return false;
         }
+        return null;
       });
 
       // when
@@ -458,10 +477,12 @@ void main() {
         'THEN should should call "resetZoom"', () async {
       // given
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'resetZoom') {
           actualMethodCall = call;
         }
+        return null;
       });
 
       // when
@@ -482,10 +503,12 @@ void main() {
       // given
       const givenZoom = 100.1;
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'setZoom') {
           actualMethodCall = call;
         }
+        return null;
       });
 
       // when
@@ -506,11 +529,13 @@ void main() {
       // given
       const givenZoom = 33.333;
       late final MethodCall actualMethodCall;
-      channel.setMockMethodCallHandler((call) async {
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (call) async {
         if (call.method == 'currentZoom') {
           actualMethodCall = call;
           return givenZoom;
         }
+        return null;
       });
 
       // when